python-backend/fixed_market_data.py
import yfinance as yf
import asyncio
import time
import logging
from datetime import datetime
import mysql.connector
from mysql.connector import Error
import os
from dotenv import load_dotenv
from typing import Dict, List
import random

load_dotenv()
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

class FixedMarketData:
    def __init__(self):
        self.db_config = {
            'host': os.getenv('DB_HOST', 'localhost'),
            'port': int(os.getenv('DB_PORT', 3306)),
            'database': os.getenv('DB_NAME', 'aimstocks'),
            'user': os.getenv('DB_USER', 'aimstocks_user'),
            'password': os.getenv('DB_PASSWORD', 'aimstocks_password')
        }
        
        self.symbols = ['RELIANCE.NS', 'TCS.NS', 'INFY.NS', 'HDFCBANK.NS', 'ICICIBANK.NS']
        
    def get_yfinance_data(self) -> Dict:
        """Get data from yfinance with proper error handling"""
        try:
            logger.info("Fetching data from yfinance...")
            
            result = {}
            
            # Fetch each symbol individually to avoid connection issues
            for symbol in self.symbols:
                try:
                    ticker = yf.Ticker(symbol)
                    
                    # Get recent data
                    hist = ticker.history(period="5d", interval="1d")
                    
                    if not hist.empty:
                        latest = hist.iloc[-1]
                        prev = hist.iloc[-2] if len(hist) > 1 else latest
                        
                        clean_symbol = symbol.replace('.NS', '')
                        
                        result[clean_symbol] = {
                            'ltp': float(latest['Close']),
                            'open': float(latest['Open']),
                            'high': float(latest['High']),
                            'low': float(latest['Low']),
                            'volume': int(latest['Volume']),
                            'change': float(latest['Close'] - prev['Close']),
                            'change_percent': float((latest['Close'] - prev['Close']) / prev['Close'] * 100)
                        }
                        
                        logger.info(f"✅ {clean_symbol}: ₹{latest['Close']:.2f}")
                        
                    # Small delay between requests
                    time.sleep(0.5)
                    
                except Exception as e:
                    logger.error(f"Error fetching {symbol}: {e}")
                    continue
            
            return result
            
        except Exception as e:
            logger.error(f"yfinance error: {e}")
            return {}
    
    def get_fallback_data(self) -> Dict:
        """Generate realistic fallback data"""
        logger.info("Using fallback data...")
        
        base_prices = {
            'RELIANCE': 2875.50,
            'TCS': 3880.25,
            'INFY': 1675.80,
            'HDFCBANK': 1595.30,
            'ICICIBANK': 1245.60
        }
        
        result = {}
        for symbol, base_price in base_prices.items():
            change_percent = random.uniform(-1.5, 1.5)
            change_value = base_price * (change_percent / 100)
            current_price = base_price + change_value
            
            result[symbol] = {
                'ltp': round(current_price, 2),
                'open': round(base_price * random.uniform(0.998, 1.002), 2),
                'high': round(current_price * random.uniform(1.0, 1.015), 2),
                'low': round(current_price * random.uniform(0.985, 1.0), 2),
                'volume': random.randint(800000, 1500000),
                'change': round(change_value, 2),
                'change_percent': round(change_percent, 2)
            }
            
            logger.info(f"📊 {symbol}: ₹{current_price:.2f} ({change_percent:+.2f}%)")
        
        return result
    
    def get_db_connection(self):
        """Get database connection"""
        try:
            return mysql.connector.connect(**self.db_config)
        except Error as e:
            logger.error(f"Database connection error: {e}")
            return None
    
    async def update_database(self, data: Dict):
        """Update database with correct schema"""
        connection = self.get_db_connection()
        if not connection:
            return False
            
        try:
            cursor = connection.cursor()
            
            for symbol, info in data.items():
                try:
                    # Fixed query without exchange column
                    query = """
                    INSERT INTO market_data (symbol, ltp, open_price, high_price, low_price, 
                                           volume, change_value, change_percent, updated_at)
                    VALUES (%s, %s, %s, %s, %s, %s, %s, %s, NOW())
                    ON DUPLICATE KEY UPDATE
                    ltp = VALUES(ltp),
                    open_price = VALUES(open_price),
                    high_price = VALUES(high_price),
                    low_price = VALUES(low_price),
                    volume = VALUES(volume),
                    change_value = VALUES(change_value),
                    change_percent = VALUES(change_percent),
                    updated_at = NOW()
                    """
                    
                    values = (
                        symbol,
                        float(info.get('ltp', 0)),
                        float(info.get('open', 0)),
                        float(info.get('high', 0)),
                        float(info.get('low', 0)),
                        int(info.get('volume', 0)),
                        float(info.get('change', 0)),
                        float(info.get('change_percent', 0))
                    )
                    
                    cursor.execute(query, values)
                    
                except Exception as e:
                    logger.error(f"Error updating {symbol}: {e}")
                    continue
            
            connection.commit()
            logger.info(f"✅ Updated {len(data)} symbols in database")
            return True
            
        except Error as e:
            logger.error(f"Database error: {e}")
            return False
        finally:
            if connection.is_connected():
                cursor.close()
                connection.close()
    
    async def fetch_and_update(self):
        """Fetch data and update database"""
        logger.info("🚀 Starting market data fetch...")
        
        # Try yfinance first
        data = self.get_yfinance_data()
        
        # Use fallback if no data
        if not data:
            data = self.get_fallback_data()
        
        # Update database
        if data:
            success = await self.update_database(data)
            if success:
                logger.info("✅ Market data update completed")
            else:
                logger.error("❌ Database update failed")
        else:
            logger.error("❌ No data available")
    
    async def run_continuous(self, interval: int = 30):
        """Run continuous updates"""
        logger.info(f"🔄 Starting continuous updates every {interval}s")
        
        while True:
            try:
                await self.fetch_and_update()
                
                # Check market hours (9:15 AM to 3:30 PM IST)
                now = datetime.now()
                if now.hour < 9 or (now.hour == 9 and now.minute < 15) or now.hour >= 16:
                    logger.info("📴 Market closed - using longer interval")
                    await asyncio.sleep(300)  # 5 minutes when market closed
                else:
                    await asyncio.sleep(interval)
                    
            except KeyboardInterrupt:
                logger.info("🛑 Stopping...")
                break
            except Exception as e:
                logger.error(f"❌ Error: {e}")
                await asyncio.sleep(60)

async def main():
    """Main function"""
    market_data = FixedMarketData()
    
    try:
        # Single update
        await market_data.fetch_and_update()
        
        # Uncomment for continuous updates
        # await market_data.run_continuous()
        
    except KeyboardInterrupt:
        logger.info("Application stopped")

if __name__ == "__main__":
    asyncio.run(main())