python-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
aiohttp==3.9.1
yfinance==0.2.28
pandas==2.1.4
numpy==1.26.2
mysql-connector-python==8.2.0
python-dotenv==1.0.0
pydantic==2.5.0
websockets==12.0
requests==2.31.0
fyers-apiv3==3.0.6